numa-design/phase-shift: fix ns/op truncation bug in primary metric
nsPerRead is an aggregate cost across all -readers parallel goroutines
(64 by default), so it is normally sub-nanosecond (~1-2 ns/read).
int64(nsPerRead) truncated every round's value down to a bare integer,
collapsing all 10 rounds per arm to exactly the same 1 or 2 and
destroying all variance in the benchstat-consumable output — the
first candidate 3 sweep run against this bug produced "all samples
are equal" for every comparison, which is not a real result. Print
with decimal precision instead.

+++ numa-design/phase-shift/main.go
@@ -175,7 +175,12 @@ func main() {
 	fmt.Fprintf(os.Stderr, "total reads=%d elapsed=%v ns/read=%.2f noop-pin(armA)=%v\n",
 		total, elapsed, nsPerRead, noopPin)
 
-	fmt.Fprintf(os.Stdout, "BenchmarkPhaseChase 1 %d ns/op\n", int64(nsPerRead))
+	// nsPerRead is the aggregate cost across all *numReaders parallel
+	// readers, so it is normally sub-nanosecond (e.g. ~1-2 ns/read with 64
+	// readers) — print with decimal precision. Truncating to an integer
+	// here previously collapsed every round to exactly 1 or 2, destroying
+	// all variance and making the metric useless to benchstat.
+	fmt.Fprintf(os.Stdout, "BenchmarkPhaseChase 1 %.4f ns/op\n", nsPerRead)
 }
 
 // allocRings allocates size bytes as node structs (64B each), linked into
